Pass the search path to cygpath only once
Calling cygpath can be slow. Previously, the search path was being
transformed every time a search was made. Additionally, just
checking for cygpath can be slow. The search path is now refactored
so that the search path from gcc is transformed just once and the
transformation step is also used to detect the presence of cygpath.