More mucking around with relocations
Except that it looks like _none_ of this is needed - go back to a
simpler flexlink without all these (park this crap elsewhere).

Essentially, all of these relocations only work for static linking -
because they rely not just on code being close, but for ldr, etc. also
on the symbol being _after_ the relocation, etc.