Parked: original 2018/2020 versions of ARM64
Some possibly useful stuff in here - refactor of parts of flexdll.c to
be slightly less unclear and some half-hearted (almost certainly broken)
initial implementations of the more complex ARM64 relocs.

The whole thing was originally attempted pre-OCaml 4.12 back before
__declspec(dllimport) was restored for Cygwin64. The switch there was
fortuitous for ARM64, because most of these relocations would require
considerable veneers and other rewriting to get the relocations to work
dynamically (if they're even possible). The 2024 version of the branch
carefully reports these ARM64 relocations by name and the expectation is
that if they're seen then it implies that there's a missing
__declspec(dllimport) somewhere in a header, rather than that the
relocation needs to be added to flexlink.